issue-1046 logger management improvements
