+++ strongbox-commons/src/main/resources/application.yaml
@@ -66,7 +66,7 @@ management:
   endpoints:
     web:
       exposure:
-        include: health,info,beans,metrics,trace,scheduledtasks,threaddump
+        include: health,info,beans,metrics,trace,scheduledtasks,threaddump,loggers
       base-path: /api/monitoring
 cacheManagerConfiguration:
   groupConfig:
