+++ strongbox-web-core/src/test/resources/application-test.yaml
@@ -1,3 +1,8 @@
 strongbox:
   basedir: ./target
   url: http://${strongbox.host:localhost}:${strongbox.port:48080}
+
+management:
+  health:
+    ldap:
+      enabled: false
